// frontend/src/types/socket.ts
export enum SocketEvent {
    JOIN_ROOM = "JOIN_ROOM",
    LEAVE_ROOM = "LEAVE_ROOM",
    SEND_MESSAGE = "SEND_MESSAGE",
    MESSAGE_RECEIVED = "MESSAGE_RECEIVED",
    USER_OFFLINE = "USER_OFFLINE",
    USER_ONLINE = "USER_ONLINE",
    ROOM_MESSAGE_SEEN = "ROOM_MESSAGE_SEEN",
    START_TYPING = "START_TYPING",
    STOP_TYPING = "STOP_TYPING",
    MESSAGE_SEEN = "MESSAGE_SEEN",
    MESSAGE_EDITED = "MESSAGE_EDITED",
    MESSAGE_DELETED = "MESSAGE_DELETED",
    EDIT_MESSAGE = "EDIT_MESSAGE",
}

export interface SocketPayload<T = any> {
    event: SocketEvent;
    data: T;
}